+++ example/lib/src/graphql/all_pokemon.req.gql.dart
@@ -5,7 +5,6 @@ import 'package:gql_example_flutter/src/graphql/all_pokemon.var.gql.dart'
     as _i3;
 import 'package:gql_exec/gql_exec.dart' as _i4;
 import 'package:gql_example_flutter/src/graphql/all_pokemon.op.gql.dart' as _i5;
-import 'package:uuid/uuid.dart' as _i6;
 
 class AllPokemon extends _i1.QueryRequest<_i2.$AllPokemon> {
   factory AllPokemon(
@@ -24,7 +23,7 @@ class AllPokemon extends _i1.QueryRequest<_i2.$AllPokemon> {
               ? buildVars(_i3.AllPokemonVarBuilder()).variables
               : const {},
           context: context != null ? context : const _i4.Context(),
-          queryId: queryId != null ? queryId : _i6.Uuid().v4(),
+          queryId: queryId,
           updateResult: updateResult,
           optimisticResponse: optimisticResponse,
           updateCacheHandlerKey: updateCacheHandlerKey,